+++ common/Cargo.toml
@@ -9,3 +9,4 @@ serde = { version = "1.0.197", features = ["derive"]}
 serde_json = "1.0.114"
 chrono = { version = "0.4.35", features = ["serde"] }
 sea-orm = "0.12"
+actix-settings = "0.7.1"
